Extend semaphore with signal features
The extension allows to link one or multiple semaphores (let's call it signal)
with another unlinked semaphore. (Note: Intentionally a signal can not be
chained with other signals for now.)

A signal can be created via an extended create_sm syscall, where the to be
linked to semaphore must be provided and a signal value. In userland the signal
is presented as a ordinary semaphore capability.

Upon up() on a signal cap leads to a normal up() in the kernel signal object as
well in the linked semaphore. If no thread was blocked the signal will be
queued up. If in the linked semaphore a thread has been
blocked, it becomes unblocked and in regs.ARG_2 and regs.ARG_3 the value as
the current counter of the signal gets delivered. Additionally the signal
counter gets reseted.

Upon dn() on a linked semaphore the calling thread becomes blocked, if the
counter is zero. If a signal was enqueued, the signal gets dequeued and the
value as the counter gets delivered and the counter in the signal gets reseted.

Upon dn() on a signal cap a BAD_CAP will be returned.

Currently the SI class is implemented as part of the SM class. Actually it
would be better the way around -> class Si: public Sm. For now I avoided that
since I would have to touch even more places of the kernel - but in the long
run this probably will be necessary anyway.